+++ src/main/java/com/meiye/bo/role/AuthUserBo.java
@@ -16,11 +16,11 @@ import java.sql.Timestamp;
  */
 @Data
 public class AuthUserBo extends ParentBo {
-	private long id;
-	private long roleId;
+	private Long id;
+	private Long roleId;
 	private AuthRoleBo roleBo;
 	private String name;
-	private byte gender;
+	private Integer gender;
 	private Timestamp birthday;
 	private String identityCard;
 	private String  education;
@@ -29,7 +29,7 @@ public class AuthUserBo extends ParentBo {
 	private String isMarry;
 	private String mobile;
 	private String email;
-	private long QQ;
+	private Long QQ;
 	private String address;
 	private String icon;
 	private String ecName;
@@ -49,13 +49,13 @@ public class AuthUserBo extends ParentBo {
 	private String account;
 	private String password;
 	private String passwordNum;
-	private byte sourceFlag;
+	private Integer sourceFlag;
 	private String salt ="";
-	private long shopIdenty;
-	private long brandIdenty;
-	private byte enabledFlag;
-	private byte assignedGroup;
-	private long assignedId;
+	private Long shopIdenty;
+	private Long brandIdenty;
+	private Integer enabledFlag;
+	private Integer assignedGroup;
+	private Long assignedId;
 
 
 }
